requirements.txt
h5py==3.7.0
ipdb==0.13.13
matplotlib==3.5.2
MedPy==0.4.0
nibabel==5.2.1
numpy==1.20.0
opencv_python==4.8.0.76
pandas==1.4.4
Pillow==9.2.0
ptflops==0.7.3
pydicom==2.4.4
pynrrd==1.0.0
scikit_image==0.19.2
scipy==1.9.1
seaborn==0.11.2
SimpleITK==2.2.1
tensorboardX==2.6.2
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.1
xlrd==2.0.1
